// packages/acp-ws-bridge/src/generated/UnsupportedVersionError.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Error returned when parsing an envelope with an unsupported version.
 */
export type UnsupportedVersionError = { 
/**
 * The version that was received.
 */
received: number, 
/**
 * The versions that are supported.
 */
supported: Array<number>, };